+++ assess/RELEASES.md
@@ -593,3 +593,13 @@ Based on:
 - [python v0.32.0] assess
 ### Releases
 - [PyPI v0.32.0] https://pypi.org/project/codat-assess/0.32.0 - assess
+
+## 2023-08-19 00:11:18
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.74.3 (2.86.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.32.1] assess
+### Releases
+- [PyPI v0.32.1] https://pypi.org/project/codat-assess/0.32.1 - assess